rename home component
