+++ app/src/styles/home.scss
